refactor: context remove, code refactor
