+++ src/components/PageLayout.tsx
@@ -1,10 +1,6 @@
 import styled from 'styled-components';
 import { Link, Outlet, useNavigate, useParams } from 'react-router-dom';
 
-// import { useAppContext } from '../context/AppContext';
-// import { useFetchData } from '../hooks/useFetchData';
-
-// import { useDispatch } from 'react-redux';
 import { useAppSelector } from '../store/store';
 import { useGetPeopleQuery } from '../store/getItemsApi';
 
@@ -71,7 +67,6 @@ const PagginationButton = styled(Link)<{ $active: boolean }>`
 export default function PageLayout() {
   const navigate = useNavigate();
   const { search_pattern, page_number } = useParams();
-  // const { currentItemList, setCurrentItemList } = useAppContext();
   const itemsPerPage = useAppSelector((state) => state.itemsPerPage.value);
   if (!page_number) {
     navigate('/page=1');
@@ -95,12 +90,6 @@ export default function PageLayout() {
       : `https://swapi.dev/api/people/?page=${Math.ceil(currentPage / 2)}`;
   }
 
-  // const { data, isLoading, isError } = useFetchData({
-  //   options: options,
-  //   currentPage: currentPage,
-  //   searchPattern: search_pattern,
-  // });
-
   const pageNumber =
     itemsPerPage === 5 ? Math.ceil(currentPage / 2) : currentPage;
 
@@ -108,8 +97,6 @@ export default function PageLayout() {
     page: pageNumber,
     search: search_pattern,
   });
-
-  console.log('data2', data);
 
   const pages = data ? Math.ceil(data.count / itemsPerPage) : 0;
   let outputArr: ICharacter[] | [] = [];
@@ -123,10 +110,6 @@ export default function PageLayout() {
       outputArr = data?.results.slice(5) as ICharacter[];
     }
   }
-
-  // if (JSON.stringify(outputArr) !== JSON.stringify(currentItemList)) {
-  //   setCurrentItemList(outputArr);
-  // }
 
   if (isFetching) {
     return <Loader />;
