+++ src/store/itemsPerPageSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
 import type { RootState } from './store';
-import type { ItemsPerPageType } from '../context/AppContext';
+import type { ItemsPerPageType } from '../types';
 
 interface ItemsPerPageStateType {
   value: ItemsPerPageType;
